modify the style of the overview page
Type: Improvement
