+++ client/applications/admin/modules/project/pop/modify_quota/quota_pop.jsx
@@ -89,17 +89,17 @@ class QuotaPop extends React.Component {
 
     types.forEach((item) => {
       quota[2].items.push({
-        title: __[item] + __.volume,
+        title: (__[item] !== undefined ? __[item] : item) + __.volume,
         key: 'volumes_' + item,
         link: 'volume'
       });
       quota[2].items.push({
-        title: __[item] + __.volume + __.gigabyte + __.unit_gb,
+        title: (__[item] !== undefined ? __[item] : item) + __.volume + __.gigabyte + __.unit_gb,
         key: 'gigabytes_' + item,
         link: 'volume'
       });
       quota[2].items.push({
-        title: __[item] + __.snapshot,
+        title: (__[item] !== undefined ? __[item] : item) + __.snapshot,
         key: 'snapshots_' + item,
         link: 'snapshot'
       });
